Add support for the M3 DS Real and derivatives
This provides support for the following:
- M3 DS Real
- iTouchDS/iTouch2
- r4rts.com carts
  - and the 1:1 clone of this cart from r4isdhc.com (black RTS)

Yet another Lua script (dsbize.lua) has been added to patch the ROM
header. On the M3, the ARM9/ARM7 offset, address, size are located
somewhere else, which is filled in by the script:
- 0xC8: ARM9 ROM offset - 0xC8
- 0xCC: ARM9 address
- 0xD0: ARM9 size
- 0xD4: ARM7 ROM offset - 0xC8
- 0xD8: ARM7 address
- 0xDC: ARM7 size

The M3 also has an XOR-encrypted header, depending on the cartridge
model and region. With XOR encryption, every byte in the header is
XORed with a key. Known keys for specific cartridges are as follows:
- M3 DS Real jp : 0x07
- M3 DS Real eng: 0x12
- M3 DS Real gb : 0x23
- iTouchDS jp   : 0x37
- iTouchDS eng  : 0x32
- iTouchDS gb   : 0x33
- r4rts.com     : 0x72

A new variant of the ARM9 has been added; this ARM9 does not implement
the bootstub struct. This is needed because the M3 firmware, as it is
using the DSBooter format, does not store the correct addresses in the
NDS header area at 0x023FFE00. Since this breaks implementing the
bootstub area, which expects these values to be correct, disable it.