dldi_patch: be more compliant with DLDI specs w/r/t patching
The DLDI area of a given ROM already provides the address of the areas
in dldiStart and dldiEnd. A few unnoticed homebrew apps, and now libnds
v2 / calico, are now writing addresses that do not actually point to
this area, but rather an expected area to be relocated to later.

Unfortunately this breaks the current DLDI patching code, because there
was an assumption that the location that the DLDI is being relocated to
will match the pointers in dldiStart and dldiEnd. Thus, use the address
provided in the ROM's area to adjust properly.